requirements.txt
discord.py==2.3.2
reportlab==4.1.0
pytz==2024.1
